MLPAB-1656 Ensure all event errors are wrapped with event name (#889)
* Ensure all event errors are wrapped with event name

* Use custom span when sending an event
