+++ cmd/event-received/cloud_watch_event_handler_test.go
@@ -1,0 +1,400 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/aws/aws-lambda-go/events"
+	"github.com/ministryofjustice/opg-modernising-lpa/internal/actor"
+	"github.com/ministryofjustice/opg-modernising-lpa/internal/date"
+	"github.com/ministryofjustice/opg-modernising-lpa/internal/dynamo"
+	"github.com/ministryofjustice/opg-modernising-lpa/internal/notify"
+	"github.com/ministryofjustice/opg-modernising-lpa/internal/page"
+	"github.com/ministryofjustice/opg-modernising-lpa/internal/uid"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+)
+
+func TestHandleUnknownEvent(t *testing.T) {
+	handler := &cloudWatchEventHandler{}
+
+	err := handler.Handle(ctx, events.CloudWatchEvent{DetailType: "some-event"})
+	assert.Equal(t, fmt.Errorf("unknown cloudwatch event"), err)
+}
+
+func TestHandleUidRequested(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "uid-requested",
+		Detail:     json.RawMessage(`{"lpaID":"an-id","donorSessionID":"donor-id","type":"hw","donor":{"name":"a donor","dob":"2000-01-02","postcode":"F1 1FF"}}`),
+	}
+
+	uidClient := newMockUidClient(t)
+	uidClient.
+		On("CreateCase", ctx, &uid.CreateCaseRequestBody{
+			Type: "hw",
+			Donor: uid.DonorDetails{
+				Name:     "a donor",
+				Dob:      date.New("2000", "01", "02"),
+				Postcode: "F1 1FF",
+			},
+		}).
+		Return("M-1111-2222-3333", nil)
+
+	uidStore := newMockUidStore(t)
+	uidStore.
+		On("Set", ctx, "an-id", "donor-id", "M-1111-2222-3333").
+		Return(nil)
+
+	err := handleUidRequested(ctx, uidStore, uidClient, event)
+	assert.Nil(t, err)
+}
+
+func TestHandleUidRequestedWhenUidClientErrors(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "uid-requested",
+		Detail:     json.RawMessage(`{"lpaID":"an-id","donorSessionID":"donor-id","type":"hw","donor":{"name":"a donor","dob":"2000-01-02","postcode":"F1 1FF"}}`),
+	}
+
+	uidClient := newMockUidClient(t)
+	uidClient.
+		On("CreateCase", ctx, mock.Anything).
+		Return("", expectedError)
+
+	err := handleUidRequested(ctx, nil, uidClient, event)
+	assert.Equal(t, fmt.Errorf("failed to create case: %w", expectedError), err)
+}
+
+func TestHandleUidRequestedWhenUidStoreErrors(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "uid-requested",
+		Detail:     json.RawMessage(`{"lpaID":"an-id","donorSessionID":"donor-id","type":"hw","donor":{"name":"a donor","dob":"2000-01-02","postcode":"F1 1FF"}}`),
+	}
+
+	uidClient := newMockUidClient(t)
+	uidClient.
+		On("CreateCase", ctx, mock.Anything).
+		Return("M-1111-2222-3333", nil)
+
+	uidStore := newMockUidStore(t)
+	uidStore.
+		On("Set", ctx, "an-id", "donor-id", "M-1111-2222-3333").
+		Return(expectedError)
+
+	err := handleUidRequested(ctx, uidStore, uidClient, event)
+	assert.Equal(t, fmt.Errorf("failed to set uid: %w", expectedError), err)
+}
+
+func TestHandleEvidenceReceived(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "evidence-required",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{PK: "LPA#123"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, map[string]string{
+			"PK": "LPA#123",
+			"SK": "#EVIDENCE_RECEIVED",
+		}).
+		Return(nil)
+
+	err := handleEvidenceReceived(ctx, client, event)
+	assert.Nil(t, err)
+}
+
+func TestHandleEvidenceReceivedWhenClientGetError(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "evidence-required",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(expectedError)
+
+	err := handleEvidenceReceived(ctx, client, event)
+	assert.Equal(t, fmt.Errorf("failed to resolve uid: %w", expectedError), err)
+}
+
+func TestHandleEvidenceReceivedWhenLpaMissingPK(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "evidence-required",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{})
+			json.Unmarshal(b, v)
+			return nil
+		})
+
+	err := handleEvidenceReceived(ctx, client, event)
+	assert.Equal(t, errors.New("PK missing from LPA in response"), err)
+}
+
+func TestHandleEvidenceReceivedWhenClientPutError(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "evidence-required",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{PK: "LPA#123"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, map[string]string{
+			"PK": "LPA#123",
+			"SK": "#EVIDENCE_RECEIVED",
+		}).
+		Return(expectedError)
+
+	err := handleEvidenceReceived(ctx, client, event)
+	assert.Equal(t, fmt.Errorf("failed to persist evidence received: %w", expectedError), err)
+}
+
+func TestHandleFeeApproved(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "reduced-fee-approved",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	now := time.Now()
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{PK: "LPA#123", SK: "#DONOR#456"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("One", ctx, "LPA#123", "#DONOR#456", mock.Anything).
+		Return(func(ctx context.Context, pk, sk string, v interface{}) error {
+			b, _ := json.Marshal(actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskPending}})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskCompleted}, UpdatedAt: now}).
+		Return(nil)
+
+	shareCodeSender := newMockShareCodeSender(t)
+	shareCodeSender.
+		On("SendCertificateProvider", ctx, notify.CertificateProviderProvideCertificatePromptEmail, page.AppData{}, &actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskCompleted}, UpdatedAt: now}).
+		Return(nil)
+
+	err := handleFeeApproved(ctx, client, event, shareCodeSender, page.AppData{}, func() time.Time { return now })
+	assert.Nil(t, err)
+}
+
+func TestHandleFeeApprovedWhenDynamoClientPutError(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "reduced-fee-approved",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	now := time.Now()
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("One", ctx, "LPA#123", "#DONOR#456", mock.Anything).
+		Return(func(ctx context.Context, pk, sk string, v interface{}) error {
+			b, _ := json.Marshal(actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskPending}})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskCompleted}, UpdatedAt: now}).
+		Return(expectedError)
+
+	err := handleFeeApproved(ctx, client, event, nil, page.AppData{}, func() time.Time { return now })
+	assert.Equal(t, fmt.Errorf("failed to update LPA task status: %w", expectedError), err)
+}
+
+func TestHandleFeeApprovedWhenShareCodeSenderError(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "reduced-fee-approved",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	now := time.Now()
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{PK: "LPA#123", SK: "#DONOR#456"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("One", ctx, "LPA#123", "#DONOR#456", mock.Anything).
+		Return(func(ctx context.Context, pk, sk string, v interface{}) error {
+			b, _ := json.Marshal(actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskPending}})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskCompleted}, UpdatedAt: now}).
+		Return(nil)
+
+	shareCodeSender := newMockShareCodeSender(t)
+	shareCodeSender.
+		On("SendCertificateProvider", ctx, notify.CertificateProviderProvideCertificatePromptEmail, page.AppData{}, &actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskCompleted}, UpdatedAt: now}).
+		Return(expectedError)
+
+	err := handleFeeApproved(ctx, client, event, shareCodeSender, page.AppData{}, func() time.Time { return now })
+	assert.Equal(t, fmt.Errorf("failed to send share code to certificate provider: %w", expectedError), err)
+}
+
+func TestHandleMoreEvidenceRequired(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "more-evidence-required",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	now := time.Now()
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{PK: "LPA#123", SK: "#DONOR#456"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("One", ctx, "LPA#123", "#DONOR#456", mock.Anything).
+		Return(func(ctx context.Context, pk, sk string, v interface{}) error {
+			b, _ := json.Marshal(actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskPending}})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskMoreEvidenceRequired}, UpdatedAt: now}).
+		Return(nil)
+
+	err := handleMoreEvidenceRequired(ctx, client, event, func() time.Time { return now })
+	assert.Nil(t, err)
+}
+
+func TestHandleMoreEvidenceRequiredWhenPutError(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "more-evidence-required",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	now := time.Now()
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{PK: "LPA#123", SK: "#DONOR#456"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("One", ctx, "LPA#123", "#DONOR#456", mock.Anything).
+		Return(func(ctx context.Context, pk, sk string, v interface{}) error {
+			b, _ := json.Marshal(actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskPending}})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskMoreEvidenceRequired}, UpdatedAt: now}).
+		Return(expectedError)
+
+	err := handleMoreEvidenceRequired(ctx, client, event, func() time.Time { return now })
+	assert.Equal(t, fmt.Errorf("failed to update LPA task status: %w", expectedError), err)
+}
+
+func TestHandleFeeDenied(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "reduced-fee-declined",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	now := time.Now()
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{PK: "LPA#123", SK: "#DONOR#456"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("One", ctx, "LPA#123", "#DONOR#456", mock.Anything).
+		Return(func(ctx context.Context, pk, sk string, v interface{}) error {
+			b, _ := json.Marshal(actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskPending}})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskDenied}, UpdatedAt: now}).
+		Return(nil)
+
+	err := handleFeeDenied(ctx, client, event, func() time.Time { return now })
+	assert.Nil(t, err)
+}
+
+func TestHandleFeeDeniedWhenPutError(t *testing.T) {
+	event := events.CloudWatchEvent{
+		DetailType: "reduced-fee-declined",
+		Detail:     json.RawMessage(`{"uid":"M-1111-2222-3333"}`),
+	}
+
+	now := time.Now()
+
+	client := newMockDynamodbClient(t)
+	client.
+		On("OneByUID", ctx, "M-1111-2222-3333", mock.Anything).
+		Return(func(ctx context.Context, uid string, v interface{}) error {
+			b, _ := json.Marshal(dynamo.Key{PK: "LPA#123", SK: "#DONOR#456"})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("One", ctx, "LPA#123", "#DONOR#456", mock.Anything).
+		Return(func(ctx context.Context, pk, sk string, v interface{}) error {
+			b, _ := json.Marshal(actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskPending}})
+			json.Unmarshal(b, v)
+			return nil
+		})
+	client.
+		On("Put", ctx, actor.DonorProvidedDetails{PK: "LPA#123", SK: "#DONOR#456", Tasks: actor.DonorTasks{PayForLpa: actor.PaymentTaskDenied}, UpdatedAt: now}).
+		Return(expectedError)
+
+	err := handleFeeDenied(ctx, client, event, func() time.Time { return now })
+	assert.Equal(t, fmt.Errorf("failed to update LPA task status: %w", expectedError), err)
+}
