+++ cmd/event-received/main.go
@@ -5,26 +5,21 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
-	"net/http"
 	"os"
 	"time"
 
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/aws/aws-sdk-go-v2/aws"
-	v4 "github.com/aws/aws-sdk-go-v2/aws/signer/v4"
 	"github.com/aws/aws-sdk-go-v2/config"
 	dynamodbtypes "github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/actor"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/app"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/dynamo"
-	"github.com/ministryofjustice/opg-modernising-lpa/internal/localize"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/notify"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/page"
-	"github.com/ministryofjustice/opg-modernising-lpa/internal/random"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/s3"
-	"github.com/ministryofjustice/opg-modernising-lpa/internal/secrets"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/uid"
 )
 
@@ -114,8 +109,6 @@ func handler(ctx context.Context, event Event) error {
 		return fmt.Errorf("failed to create dynamodb client: %w", err)
 	}
 
-	now := time.Now
-
 	if event.isS3Event() {
 		s3Client := s3.NewClient(cfg, evidenceBucketName)
 		documentStore := app.NewDocumentStore(dynamoClient, nil, nil, nil, nil)
@@ -128,56 +121,21 @@ func handler(ctx context.Context, event Event) error {
 	}
 
 	if event.isCloudWatchEvent() {
-		err := fmt.Errorf("unknown cloudwatch event")
-
-		switch event.DetailType {
-		case "uid-requested":
-			uidStore := app.NewUidStore(dynamoClient, now)
-			uidClient := uid.New(uidBaseURL, &http.Client{Timeout: 10 * time.Second}, cfg, v4.NewSigner(), time.Now)
-
-			err = handleUidRequested(ctx, uidStore, uidClient, event.CloudWatchEvent)
-
-		case "evidence-received":
-			err = handleEvidenceReceived(ctx, dynamoClient, event.CloudWatchEvent)
-
-		case "reduced-fee-approved":
-			bundle := localize.NewBundle("./lang/en.json", "./lang/cy.json")
-
-			//TODO do this in handleFeeApproved when/if we save lang preference in LPA
-			appData := page.AppData{Localizer: bundle.For(localize.En)}
-
-			secretsClient, err := secrets.NewClient(cfg, time.Hour)
-			if err != nil {
-				return fmt.Errorf("failed to create secrets client: %w", err)
-			}
-
-			notifyApiKey, err := secretsClient.Secret(ctx, secrets.GovUkNotify)
-			if err != nil {
-				return fmt.Errorf("failed to get notify API secret: %w", err)
-			}
-
-			notifyClient, err := notify.New(notifyIsProduction, notifyBaseURL, notifyApiKey, http.DefaultClient)
-			if err != nil {
-				return err
-			}
-
-			shareCodeSender := page.NewShareCodeSender(app.NewShareCodeStore(dynamoClient), notifyClient, appPublicURL, random.String)
-
-			err = handleFeeApproved(ctx, dynamoClient, event.CloudWatchEvent, shareCodeSender, appData, now)
-
-		case "reduced-fee-declined":
-			err = handleFeeDenied(ctx, dynamoClient, event.CloudWatchEvent, now)
-
-		case "move-evidence-required":
-			err = handleMoreEvidenceRequired(ctx, dynamoClient, event.CloudWatchEvent, now)
+		handler := &cloudWatchEventHandler{
+			dynamoClient:       dynamoClient,
+			now:                time.Now,
+			uidBaseURL:         uidBaseURL,
+			cfg:                cfg,
+			notifyIsProduction: notifyIsProduction,
+			notifyBaseURL:      notifyBaseURL,
+			appPublicURL:       appPublicURL,
 		}
 
-		if err != nil {
+		if err := handler.Handle(ctx, event.CloudWatchEvent); err != nil {
 			return fmt.Errorf("%s: %w", event.DetailType, err)
 		}
 
 		log.Println("successfully handled ", event.DetailType)
-
 		return nil
 	}
